+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 - [KOGITO-3793](https://issues.redhat.com/browse/KOGITO-3793) - Review OLM UI spec descriptors
 - [KOGITO-5834](https://issues.redhat.com/browse/KOGITO-5834) - Change project structure to multi-group
 - [KOGITO-5760](https://issues.redhat.com/browse/KOGITO-5760) - Upgrade operator-sdk in Kogito-operator
+- [KOGITO-5910](https://issues.redhat.com/browse/KOGITO-5910) - Unify kogito community and product into one
 - [KOGITO-5976](https://issues.redhat.com/browse/KOGITO-5976) - Kogito operator de-scoping - support for AllNamespaces
 
 ## Bug Fixes